+++ RecipeManagement.Domain/Entities/Models/Recipe.cs
@@ -15,7 +15,7 @@ namespace RecipeManagement.Domain.Entities.Models
 
         [MaxLength(40)]
         public required string Author { get; set; }
-        public required short Rating { get; set; }
+        public required float Rating { get; set; }
 
     }
 }
